[20190416] 커스텀셀렉션 고도화 소스 커밋 1. 대용량 챕터 폰트크기 키운 경우 페이징 이슈 수정 2. fixedlayout 마지막 페이지 노티 안가는 이슈 수정

+++ viewersdk/src/main/java/com/ebook/epub/viewer/EPubViewer.java
@@ -2317,7 +2317,7 @@ public class EPubViewer extends ViewerBase {
         if( mHtmlDirectionType == 1 )
             l = (mTotalPageInChapter * this.getWidth()) - l - this.getWidth();
 
-        int currentPage = (int) Math.round((double)l / (double)this.getWidth());
+        int currentPage = Math.round(l / this.getWidth());
 
         if( __requestPageMove ){
             __requestPageMove = false;
